requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.4
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-registration==3.1.2
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django-webpack-loader==1.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docxcompose==1.3.1
docxtpl==0.11.4
future==0.18.2
html5lib==1.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
numpy==1.20.1
oauthlib==3.1.0
pandas==1.1.5
Pillow==8.1.2
pycparser==2.20
PyJWT==2.1.0
pylokit==0.8.1
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.17.0
python3-openid==3.2.0
pytz==2021.1
pyuploadcare==2.7.0
reportlab==3.5.66
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stripe==2.56.0
templateddocs==0.3.4
tinycss2==1.1.0
uritemplate==3.0.1
urllib3==1.26.4
WeasyPrint==52.4
webencodings==0.5.1
xhtml2pdf==0.2.5
